+++ app/page.tsx
@@ -3,7 +3,7 @@ import { WorldMap } from "@/components/map/world-map";
 
 export default function Home() {
   return (
-    <main className="flex min-h-screen flex-col px-6 pt-12 pb-16">
+    <main className="flex flex-1 flex-col px-6 pt-12 pb-16">
       <div className="mx-auto flex w-full max-w-6xl flex-1 flex-col">
         <HeroText />
         <WorldMap />